requirements_gpu.txt
torch>=1.2
scipy
numpy
scikit-learn
gitpython
pandas
biopython
pillow
